r91: Switch to using docbook2man instead of docbook-to-man, because the former generates roff output which is actually syntactically correct. On the plus side, this will also allow some distributors to build it with tools they actually distribute (hi Gentoo!).
While I'm at it, move nbd-client.8 and nbd-server.1 to both EXTRA_DIST and
MAINTAINERCLEANFILES, so that they are now in the .tar.{bz2,gz} but still get
rebuilt when changed. This way, distributors don't have to install a whole SGML
toolkit to build the nbd utils if they don't want to (hi FreeBSD!).