Change meaning of -swap option
Change meaning of -swap option: rather than trying to use a kernel patch
which has been effectively dead since sometime around 2.1.something, do
an mlockall() so that the kernel won't throw us out.

The MCL_CURRENT | MCL_FUTURE bit might prove to be problematic, but if
so, we'll deal with it when we get there.