Remove support for timeout option
The timeout option has been broken for a very long time. Additionally,
something like this is better done by making use of TCP keepalive
probes, which nbd-server has implemented since 2003 (unless you're using
it in inetd mode, in which case you should add the keepalive option to
the inetd configuration for the nbd-server).