eth0 unalterably hardcoded in initramfs script
As it stands, and as mentioned in the README.Debian which points to bug
 #460569 which has been closed, it will currently only work if the
nbd-server is connected to a working eth0.

initramfs-tools closed their bug as they now export IP from init (and
IPOPTS is no more).   As best I understand all the implications this
means that the nbd initramfs-tools script could remove all IP handling
as long as it is run by an IP exporting init.  To cover it's arbitrary
use outside a normal initramfs-tools it could still replicate gathering
and setting IP from the kernel's ip command line.

Like this if you boot without setting ip it seems that klibc ipconfig
will complain and fail so in the patch I've included below I set
IP='eth0' in the abscence of a given ip value which maintains the
existing default behaviour unless or until configure-networking seems to
work sanely without any IP input.

Thanks again for your work on the nbdtools in debian and hope this
helps,