initrd 3 argument nbdroot sed function can fail
in a "live" busybox+klibc based initramfs the sed function to extract
the port from the 3 argument form of nbdroot fails.   The parts of the
regex work but the total regexp fails to do anything, I didn't really
dig into the source of that issue as it seemed like it was likely to be
a limitation of the particular sed.   I'm hoping you would consider
rewriting this to make it less likely to break in less competent
implementations of sed.

I have attached a patch which eliminates sed completely by using more
shell substitutions, though just breaking down that one more complicated
sed expression to two seperate calls to sed works equally well for me in
my environment so tmtowtdi.

It's worth noting that a kernel root option will set the nbd device
which is what is additionally set in the 3 argument form of nbdroot
anyway, so again I might only have tripped this as I'm not using
initramfs-tools local with a normal rootfs so I don't want to set root
on the kernel command line.

As far as I know this is only an inconvience to me though as in my case
I think I can safely set e.g. root=/dev/nbd0, I just don't want to and
the existing code teased me into thinking I wouldn't need to.

(fixed one minor error in the patch myself -- Wouter)