Fix 'make distcheck'
if we want the test suite to remain working, it has to have all the data
it wants. Ship it with the .tgz so the test suite doesn't have to
generate it (though I suppose we could make it do so if we wanted).

Also, rename the file so it's clear what it's needed for, and so it
isn't accidentally overwritten just because it's the default name.