nbd-server reports 'negotiation failed' after 'nbd-client -l', and fails on next connection
When a client requests an export list from the server,
the latter issues the message 'Error: negotiation failed'.
This is rather confusing, as the only 'problem' is (AFAICT)
that the negotiation phase did everything.

On a subsequent attempt to connect a block device, the server
fails altogether:
Client output:
	--------------------
	nasc0:root ~ 216 # /tmp/nbd-client.master -l nass0
	Negotiation: ..
	nasc0-nbd0
	nasc0:root ~ 217 # /tmp/nbd-client.master nass0 -N nasc0-nbd0 /dev/nbd0
	Negotiation: ..size = 856941734MBError: Exported device is too big for me. Get 64-bit machine :-(

	Exiting.
	--------------------
Server output:
	--------------------
	nass0:root ~ 135 # /tmp/nbd-server.master -d -C /etc/nbd-server/config
	Error: negotiation failed
	Exiting.
	Error: Read failed: Connection reset by peer
	Exiting.
	--------------------
From a tcpdump, it can be observed that the syslog messages end up
on the client socket, which makes the client choke, naturally.

I tested this for the latest git version, and the same problem
occurs. I have traced the problem to the fact that negotiate(),
in two locations, closes the socket, and returns NULL , after which
serveloop also closes it, actually closing the syslog socket instead.
The next accept() returns that same filedescriptor, which syslog()
still thinks it owns.

Note: I added the message 'Session terminated by client', because
the 'ABORT' might also be intended to be used as an irregular
end-of-session (in the future?), besides indicating a regular
end-of-session after a LIST.

--------------------------------------------