nbd-client: Change init script and config file to better deal with NBD_NAME
The config file change's purpose is mostly documenting the existence
of NBD_NAME, and the fact NBD_PORT is optional with NBD_NAME.

The init script change allows NBD_PORT to be unspecified (i.e. default),
and complains if both NBD_PORT and NBD_NAME are missing.
Without this change, the init script fails if NBD_PORT is empty or unset.