Clean up cliserv.h
Previously, cliserv.h also contained some functions, which would then be
defined twice if we hadn't inlined them.

That is a fairly ugly way of doing things, so instead move the few
functions in cliserv.h to a new cliserv.c, and compile and link that in.

This should also make more recent versions of clang not be unhappy about
"missing" functions