Fix up parse_cfile signature
parse_cfile repurposed the genconf parameter as a boolean flag (in
addition to its storing of generic configuration) to signal the
expectation (or absense thereof) of a [generic] section in the
configuration file.

Not only is this confusing, it also makes it impossible to reliably
detect whether we want an oldstyle export. Fix for that in the next
commit.