gznbd: use atoll() to convert size argument
gznbd works incorrectly with large size argument which overflows 32-bit integer.

For example:
gznbd /dev/nbd0 /root/test.gz 1000204886016
gznbd: 1000204886016 does not appear to be a valid size

The size is stored in u64, but is converted from a string using atol().
Simply replace atol() with atoll().
gznbd is compiled with -std=gnu99 (as the whole nbd), so atoll() is available.

Links to the original patch:
[1] https://sourceforge.net/p/nbd/mailman/nbd-general/thread/560411311771602%40web126.yandex.ru/#msg27860052
[2] http://marc.info/?l=nbd-general&m=131177162001072&w=2