Drop support for the oldstyle protocol
The newstyle protocol was originally written for NBD 2.9.17, released in
2009. We are 2015 now. People have had over five years to migrate their
setups now.

Additionally, due to the fact that I originally wrote support for the
newstyle protocol in a somewhat confusing way, maintaining that part of
the code has become harder than it needs be. A bug did now appear
because I couldn't read my own code anymore.

While it should, in theory, be possible to just fix the damn code while
still retaining support for the oldstyle protocol, it's easier (and not
too unreasonable) to just drop that instead.

Also fix the test suite so it doesn't try to run tests using the oldstyle
protocol anymore