Stop offering to convert old pre-2.9 config files
NBD 2.9, released in 2007, was the first NBD release which would parse a
configuration file for nbd-server. It made sense, at first, to convert
the shell script snippet "configuration files" that we supported then
into proper nbd-server config files, but that isn't useful anymore, so
drop that.

If a user still has an etch (!) or older machine lying around for which this
code was meant, they can always upgrade to something more recent first...