define error values as part of the protocol
Right now, NBD includes potentially platform-specific error values in
the protocol.

Luckily, most common error values are more or less universal: of all
errno values <= 34 (up to ERANGE), only 11 seems to differ across the
common platforms: it is EAGAIN on Windows and Linux, but it is EDEADLK
on Darwin and the *BSDs.

This patch defines a limited set of errno values that are valid for the
NBD protocol, and specifies recommendations for what error to return
in specific corner cases.  The set of errno values is roughly based on
the errors listed in the read(2) and write(2) man pages, with some
exception:

- ENOMEM is added for servers that implement copy-on-write or other
  formats that require dynamic allocation.

- EDQUOT is not part of the universal set of errors; it can be changed
  to ENOSPC on the wire format.

- EFBIG is part of the universal set of errors, but it is also changed
  to ENOSPC because it is pretty similar to ENOSPC or EDQUOT.

The server is changed to match the recommendations.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Wouter Verhelst <w@uter.be>