Remove file descriptor state dependency
We used to do lseek() followed immediately by read() or write(). This
works, but depends on the current position of the file descriptor.

Since we want to move to a multithreaded implementation, replace the two
calls by pread()/pwrite() instead, which combine the lseek() and
read()/write() calls in one, removing the file descriptor offset
dependency.