Invert return logic
The READ_ALL_ERRCHK() does not touch the "retval" variable, so if it
fails it will goto some error handling label which returns the existing
value of retval, which is initialized to 0, for "no error". As a result,
if we encounter an error in read(), currently the test will return
success, which is wrong.

We don't want to do this. Instead, make the default return value be -1,
and only set it to 0 when we reach the end of the test.