Improve formatting of numbers
When running "make check", if the number of requests is increased by an
order of magnitude and then reduced again, the old number remains
on-screen. This is confusing, since it makes the user believe that the
numbers have gone up.

So, fix that by adding two spaces after the output, which should
overwrite previous numbers.