Reduce sleep time
Currently, when we fork, we immediately try to trigger the partition
scanning, which will fail, and then we wait a whole second (which is way
longer than most connection attempts will take). This unneedlessly makes
it take a full second for the setup to be done.

Instead, turn this into a nanosleep which waits for .1 second. This is
probably still too long, but since we need to poll for this, making it
shorter is probably a bad idea.