Clean out oldstyle support
When removing oldstyle in a hurry back when 3.11 was released, all code
paths that lead to "the SERVER type has stuff set for oldstyle sockets"
were removed, but the code paths that actually checked for those sockets
and mangled with them weren't. As such, there was still a lot of stuff
there, stuff which would confuse the code and make it less easy to read.

So, remove.

While at it, we also removed two functions that dealt with SERVER types:
- `append_serve`: this would connect a server and add it to a array of
  servers. After removing everything that was specific to oldstyle
  exports, all that remained was a single call to g_array_append_val,
  for which we don't need a separate function.
  Also remove the `append` test from the test suite, which did nothing
  beyond testing the functionality of `append_serve`
- `setup_serve`: This would create the socket of a server and configure
  it (setsockopt() et al). This function as oldstyle-specific (the
  modernstyle sockets were configured elsewhere), so was removed in its
  entirety.

After all that code was removed, we could also remove the `socket`,
`socket_family` and `port` members of the SERVER type.