Properly handle the IPv4/IPv6 case
What asking for "anything" using AF_UNSPEC does is horribly ill defined.
It used to be the case that Linux would return something that would
listen to both IPv4 and IPv6 on the same socket, but this is no longer
the case.

As such, just stop pretending and loop over multiple listening
addresses, instead, defaulting to "::, 0.0.0.0".

As a side benefit, this allows people to specify multiple addresses to
listen to.