Fix TRIM implementation
Previously, we would happily try to trim when readonly but not
treefiles, which is broken; we would happily trim the original file when
copy-on-write was on, which is equally broken; and we would happily trim
beyond the end of the device.

Fix all three cases.

Additionally, the multiple file trim was broken because we didn't
increment a counter. Fix as well.