Properly get the declarations of pread/pwrite
When compiling 3.12.1, warnings will occur with '-Wall':

    nbd-server.c: In function 'rawexpwrite':
    nbd-server.c:959: warning: implicit declaration of function 'pwrite'
    nbd-server.c: In function 'rawexpread':
    nbd-server.c:1054: warning: implicit declaration of function 'pread'

In pread(2), pread/pwrite has the following Feature Test Macro Requirements:

    pread(), pwrite():
              _XOPEN_SOURCE >= 500
                        || /* Since glibc 2.12: */ _POSIX_C_SOURCE >= 200809L

So we have to define _XOPEN_SOURCE to actually get pread/pwrite. And without
the declarations, we will not get the Large File Support for pread/pwrite.

And if _XOPEN_SOURCE is defined, we have to explicitly define _BSD_SOURCE to
get the DT_* macros.