Remove this symlink
The upstream build system still refers to the "old" name of
doc/proto.md, and as such ships proto.txt as a file rather than a
symlink. This causes dpkg-source to panic over the fact that it can't
represent that in the diff.

Work around by just removing the symlink. It's not shipped in any
package, anyway.