Move include statement to the correct file.
We need the <sys/mount.h> include file just for the BLKGETSIZE64 ioctl()
definition, which has moved to nbdsrv.c. We were still doing that
include in nbd-server.c, but not in nbdsrv.c; this turned the bit that
tries to test for the size of a device using an ioctl into dead code.

Fix.