Don't reuse unsigned int variable
We want to store the return value of the read() system call in
something, and thought it might be a good idea to store it in len.

However, len is an unsigned variable, and read() might return a negative
value. This gets horrible array overflows if read does return an
error...

Coverity CID#1346480