Fix race condition
open_treefile() tries to open a file. If it doesn't exist yet, it will
create it, and then ensure that the file is of the correct length (by
lseek()ing to the end of the file and writing a single byte), so that
subsequent read()s will work:

  open() // fails
    open(O_CREAT)
    lseek()
    write()

Unfortunately, that doesn't work if we have multiple threads trying to
open the same file at the same time:

  T1                          T2
  open() // fails
    open(O_CREAT)
                              open() // succeeds
			      // read request handling will now fail
			      // (file too small); write request
			      // handling will succeed for this thread.
    lseek()
    write() // oops, we just
            // lost a byte

Fix by reusing the client connection's mutex. Ideally we wouldn't do
that and try to get this to be done atomically instead, but for now this
will have to do.