nbd: FTBFS on hurd-i386 due to getpeername on unix socket
nbd FTBFS on hurd-i386 because the unix test fails while calling
getpeername on the unix socket. getpeername is indeed not supported on
GNU/Hurd because it does not really make sense: the peer does not have
an address.  The attached patch fixes this by calling getsockname (which
does make sense) first to determine whether this is a unix or tcp/ip
socket, and then call getpeername in the tcp/ip case.