Don't hardcode values that shouldn't be hardcoded
The size of addr.sun_path isn't defined. It happens to be 107 bytes on
most Linux systems, but this is not the correct value on kFreeBSD, where
it happens to be 104 bytes instead.

Rather than hardcoding this value, use the sizeof operator to figure out
the size of the sun_path member at compile time.