doc: Clean up wording on UTF-8 and REP_SERVER usage
Consistently use the spelling UTF-8.

Make it clear that when a UTF-8 string is transmitted, the sender
need not transmit a terminating NUL byte, so the receiver must pay
attention to length.  Require error messages to be UTF-8 (thankfully,
this would include the case of an all-ASCII LC_ALL=C strerror
message).

Clean up the wording related to NBD_REP_SERVER when there is extra
data beyond the name, calling out specifically that the extra data
is either a UTF-8 string (with no further definition), or
a specific struct defined by the particular client request.  Fix
the formatting (the note about SELECT changing the semantics of
the extra data is not a field within the data).

Signed-off-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Wouter Verhelst <w@uter.be>