Don't error on NBD_FLAG_CMD_FUA on non-writes
Commit ab22e082 changed the server to reject all unknown
flags with EINVAL, but in the process broke qemu as a client
(which in qemu 2.5 unconditionally sets FUA on NBD_CMD_FLUSH).
Commit e82baa5f relaxed the protocol to explicitly document
that servers must not reject FUA on any command (but MAY
ignore it where the command does not actually cause a write).
Fix the reference implementation to match.

Signed-off-by: Eric Blake <eblake@redhat.com>
Reviewed-by: Alex Bligh <alex@alex.org.uk>
Signed-off-by: Wouter Verhelst <w@uter.be>