Fix whitespacing and indentation in nbd-tester-client.c
I realise whitespace changes are generally frowned on, but nbd-tester-client.c
is almost impossible to edit in (e.g.) emacs even in 'linux' mode, because
its whitespacing is inconsistent and does not follow the style of the rest
of the package.

This is a whitespace ONLY change and is the equivalent of running Lindent.
Lindent does:
  indent -npro -kr -i8 -ts8 -sob -l80 -ss -ncs -cp1 -il0 nbd-tester-client.c
on indent >= v2.2.10, and
  indent -npro -kr -i8 -ts8 -sob -l80 -ss -ncs -cp1 nbd-tester-client.c
otherwise. See:
  https://github.com/torvalds/linux/blob/master/scripts/Lindent

Signed-off-by: Alex Bligh <alex@alex.org.uk>
Signed-off-by: Wouter Verhelst <w@uter.be>