doc: Use sequence of replies for NBD_OPT_INFO/GO
Since NBD_OPT_INFO and NBD_OPT_GO are experimental, we still have
a chance to fix them up before promoting them to stable.

Attempting to reuse NBD_OPT_SERVER as the reply to NBD_OPT_INFO and
NBD_OPT_GO has a few problems: clients must be prepared to parse
two different styles of the reply, based on which option request
the reply is answering (not impossible, but awkward).  Extending
the information to provide even more details, like block sizing, is
likewise awkward (the only way to do it within a single reply is to
have multiple length fields that must all be consistent; but
pre-computing the overall header length may be difficult).  And
requiring the server to parrot back the export name is wasteful if
the client's name is already in canonical form.

Solve this by instead making the valid response be a series of the
new NBD_REP_INFO reply messages, followed by a concluding NBD_REP_ACK
for success or NBD_REP_ERR_* for an appropriate error (similar to how
NBD_OPT_LIST has a series).  On success, the series must include
NBD_INFO_EXPORT, so that the server provides at least as much
information as it did for NBD_OPT_EXPORT_NAME.  But prior to the
terminal reply message, the server can now send as many additional
optional items of information as it wants (clients must ignore the
ones they don't recognize); this patch starts with the canonical
name and description of the export (which is all the more
NBD_REP_SERVER was previously being used for).  A future patch
will then add another piece of information, for advertising server
block sizes.

Additionally:

- The wording about repeated replies to OPT_INFO followed by OPT_GO
was a bit repetitive; say the same thing in fewer sentences.

- Swap paragraph ordering so that NBD_REP_INFO details aren't
split by a side-note about NBD_REP_ERR_UNSUP.

- Make it clear that NBD_OPT_LIST can end in an error, such as
NBD_REP_ERR_SHUTDOWN

Signed-off-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Alex Bligh <alex@alex.org.uk>